+++ src/coniks.rs
@@ -690,7 +690,7 @@ pub fn verify_absence(
     }
 }
 
-#[cfg(test)]
+#[cfg(all(test, not(target_arch = "wasm32")))]
 mod tests {
     use super::*;
     use crate::vrf::Ecvrf;