+++ src/policy.rs
@@ -1117,7 +1117,7 @@ impl<'a> Cursor<'a> {
     }
 }
 
-#[cfg(test)]
+#[cfg(all(test, not(target_arch = "wasm32")))]
 mod tests {
     use super::*;